+++ resources/js/usePerformanceTracker.svelte.js
@@ -6,7 +6,7 @@
  * data to Svelte components.
  *
  * @package ZeroBoiler Analytics
- * @version 263.0.0
+ * @version 264.0.0
  */
 
 import { writable, derived } from 'svelte/store';